requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.2.0
autopep8==2.0.2
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.6.2
Brotli==1.0.9
build==1.2.1
cachetools==5.3.3
celery==5.3.5
certifi==2023.5.7
cffi==1.16.0
cfgv==3.3.1
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
ConfigArgParse==1.5.3
coverage==7.2.7
cryptography==41.0.4
dataclasses-json==0.6.5
Deprecated==1.2.14
distlib==0.3.6
distro==1.9.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.12.1
flake8==6.1.0
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flatbuffers==24.3.25
flower==1.2.0
frozenlist==1.4.1
fsspec==2024.3.1
gevent==22.10.2
geventhttpclient==2.0.9
google-auth==2.29.0
googleapis-common-protos==1.59.1
greenlet==3.0.3
grpcio==1.62.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
humanize==4.6.0
identify==2.5.24
idna==3.4
importlib-metadata==6.0.1
importlib_resources==6.4.0
inflect==7.0.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
jdatetime==4.1.1
Jinja2==3.1.2
jose==1.0.0
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.3
kubernetes==29.0.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-experimental==0.0.55
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.52
locust==2.15.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.1
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.7
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-propagator-b3==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.10.2
overrides==7.7.0
packaging==23.2
passlib==1.7.4
platformdirs==3.5.3
pluggy==1.0.0
posthog==3.5.0
pre-commit==3.3.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
psycopg2-binary==2.9.9
pulsar-client==3.5.0
pyasn1==0.5.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
Pygments==2.17.2
pypdf==4.1.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysqlite3-binary==0.5.2.post3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-logging-loki==0.3.1
python-multipart==0.0.9
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
redis==4.5.5
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339==6.2
rich==13.7.1
roundrobin==0.0.4
rsa==4.9
sentry-sdk==1.34.0
shellingham==1.5.4
six==1.16.0
smsir-python==1.0.7
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlalchemy-cockroachdb==2.0.1
sqlparse==0.4.4
sqltap==0.3.11
starlette==0.37.2
suds==1.1.2
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
tiktoken==0.6.0
tokenizers==0.19.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.66.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.10.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.19.0
vine==5.1.0
virtualenv==20.23.0
watchfiles==0.21.0
wcwidth==0.2.6
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.4
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
